solution/requirements.txt
aioredis==2.0.1
asyncpg==0.30.0
fastapi==0.115.8
greenlet==3.1.1
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.38
uvicorn==0.34.0
g4f[all]
